statement-2/requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
decorator==5.2.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai-whisper==20250625
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
proglog==0.1.12
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
setuptools==82.0.1
shellingham==1.5.4
sympy==1.14.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
yt-dlp==2026.3.13
